feat: 모달 닫기 기능 구현 및 리팩토링
