+++ src/js/utils/dom.ts
@@ -1,5 +1,7 @@
-export const $ = (selector: string, node = document): HTMLElement =>
-  node.querySelector(selector) as HTMLElement;
+export const $ = (
+  selector: string,
+  node: HTMLElement | Document = document
+): HTMLElement => node.querySelector(selector) as HTMLElement;
 
 export const $$ = (selector: string, node = document): NodeList =>
   node.querySelectorAll(selector) as NodeList;
